refactor(docs): share the chat runtime construction across the four surfaces
each of the four providers rebuilt the same transport, and two of them also rebuilt the same anonymous cloud, so the plumbing lived four times over while only the composition differed. useDocsChatRuntime now owns the shared part and the providers are thin.

the four are not collapsed into one component: they differ on cloud, endpoint, adapters, toolkit, suggestions and devtools, and a single component covering all four would be a configuration surface with one caller per combination.

speech and dictation adapters carry per-utterance state, which the docs provider handled and the artifacts and playground providers did not, rebuilding them on every render. the shared helper memoizes them for all three.

the directory is renamed because it held no React contexts, only runtime providers; the real ones live under components/pages/docs/contexts. that also retires a file named AssistantRuntimeProvider whose export was DocsAssistantRuntimeProvider and which shadowed the upstream symbol it imported.
