+++ apps/docs/components/pages/docs/layout/docs-root-layout.tsx
@@ -14,8 +14,10 @@ import {
   DocsAssistantPanel,
   DocsShell,
 } from "@/components/pages/docs/layout/docs-layout";
-import { DocsRuntimeProvider } from "@/contexts/DocsRuntimeProvider";
-import { DocsAssistantRuntimeProvider } from "@/contexts/AssistantRuntimeProvider";
+import {
+  DocsAssistantRuntimeProvider,
+  DocsRuntimeProvider,
+} from "@/runtimes/providers";
 import { CurrentPageProvider } from "@/components/pages/docs/contexts/current-page";
 import { PlatformProvider } from "@/components/pages/docs/platform/context";
 import { FloatingComposer } from "@/components/pages/docs/assistant/floating-composer";
