--- apps/docs/contexts/DocsRuntimeProvider.tsx
@@ -1,86 +1,0 @@
-"use client";
-
-import { useMemo } from "react";
-import {
-  AssistantRuntimeProvider,
-  WebSpeechSynthesisAdapter,
-  WebSpeechDictationAdapter,
-  CloudFileAttachmentAdapter,
-  AssistantCloud,
-  useAui,
-  Tools,
-  Suggestions,
-  unstable_Interactables,
-} from "@assistant-ui/react";
-import { AssistantChatTransport, useChatRuntime } from "@assistant-ui/ai-sdk";
-import { DevToolsModal } from "@assistant-ui/react-devtools";
-import { lastAssistantMessageIsCompleteWithToolCalls } from "ai";
-import { anonymousSessionFetch } from "@/lib/anonymous-session-client";
-import { feedbackAdapter } from "@/lib/feedback-adapter";
-import docsToolkit from "@/lib/docs-toolkit";
-
-export function DocsRuntimeProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const assistantCloud = useMemo(
-    () =>
-      new AssistantCloud({
-        baseUrl: process.env.NEXT_PUBLIC_ASSISTANT_BASE_URL!,
-        anonymous: true,
-      }),
-    [],
-  );
-
-  // Speech/dictation adapters keep internal state; create per component instance.
-  const adapters = useMemo(
-    () => ({
-      speech: new WebSpeechSynthesisAdapter(),
-      dictation: new WebSpeechDictationAdapter(),
-      feedback: feedbackAdapter,
-      attachments: new CloudFileAttachmentAdapter(assistantCloud),
-    }),
-    [assistantCloud],
-  );
-
-  const runtime = useChatRuntime({
-    transport: new AssistantChatTransport({
-      fetch: anonymousSessionFetch,
-    }),
-    sendAutomaticallyWhen: lastAssistantMessageIsCompleteWithToolCalls,
-    adapters,
-    cloud: assistantCloud,
-  });
-
-  const aui = useAui({
-    tools: Tools({ toolkit: docsToolkit }),
-    unstable_interactables: unstable_Interactables(),
-    suggestions: Suggestions([
-      {
-        title: "What's the weather",
-        label: "in San Francisco?",
-        prompt: "What's the weather in San Francisco?",
-      },
-      {
-        title: "Explain React hooks",
-        label: "like useState and useEffect",
-        prompt: "Explain React hooks like useState and useEffect",
-      },
-      {
-        title: "Show a live dashboard",
-        label: "with the present tool",
-        prompt:
-          "Use the present tool to show a compact sales dashboard: a Card with two Facts in a Row and a bar Chart of monthly sales.",
-      },
-    ]),
-  });
-
-  return (
-    <AssistantRuntimeProvider aui={aui} runtime={runtime}>
-      {children}
-
-      <DevToolsModal />
-    </AssistantRuntimeProvider>
-  );
-}
