--- apps/docs/contexts/PlaygroundRuntimeProvider.tsx
@@ -1,29 +1,0 @@
-"use client";
-
-import {
-  AssistantRuntimeProvider,
-  WebSpeechSynthesisAdapter,
-} from "@assistant-ui/react";
-import { AssistantChatTransport, useChatRuntime } from "@assistant-ui/ai-sdk";
-import { anonymousSessionFetch } from "@/lib/anonymous-session-client";
-
-export function PlaygroundRuntimeProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const runtime = useChatRuntime({
-    transport: new AssistantChatTransport({
-      fetch: anonymousSessionFetch,
-    }),
-    adapters: {
-      speech: new WebSpeechSynthesisAdapter(),
-    },
-  });
-
-  return (
-    <AssistantRuntimeProvider runtime={runtime}>
-      {children}
-    </AssistantRuntimeProvider>
-  );
-}
